Update ggml-metal.m
The reason to use Objective-C here is because you're interfacing with Metal, which is an Apple API primarily designed to be used with Objective-C or Swift. While it is technically possible to use C++ to interact with Objective-C code (in a language blend known as Objective-C++), it requires careful management of the language interop and may not offer any clear advantages in this particular case.